@@ -1,33 +1,76 @@
-"""
-See https://www.sphinx-doc.org/en/master/usage/configuration.html
-"""
-from datetime import datetime
+"""See https://www.sphinx-doc.org/en/master/usage/configuration.html."""
+from datetime import datetime, timezone
 from operator import itemgetter
-from os.path import dirname
-from typing import Final
-import os
+from pathlib import Path
 import sys
 
 import tomlkit
 
-with open(f'{dirname(__file__)}/../pyproject.toml') as f:
-    poetry = tomlkit.load(f).unwrap()['tool']['poetry']
-    authors, name, version = itemgetter('authors', 'name', 'version')(poetry)
+with (Path(__file__).parent.parent / 'pyproject.toml').open(newline='\n', encoding='utf-8') as f:
+    project_ = tomlkit.load(f).unwrap()['project']
+    authors_list, name, version = itemgetter('authors', 'name', 'version')(project_)
+authors = [f'{d["name"]} <{d["email"]}>' for d in authors_list]
 # region Path setup
 # If extensions (or modules to document with autodoc) are in another directory, add these
 # directories to sys.path here. If the directory is relative to the documentation root, use
-# os.path.abspath to make it absolute, like shown here.
-sys.path.insert(0, os.path.abspath('..'))
+# str(Path().parent.parent) to make it absolute, like shown here.
+sys.path.insert(0, str(Path(__file__).parent.parent))
 # endregion
-author: Final[str] = authors[0]
-copyright: Final[str] = str(datetime.now().year)
-project: Final[str] = name
-version: Final[str] = version
-release: Final[str] = f'v{version}'
-extensions: Final[list[str]] = (['sphinx.ext.autodoc', 'sphinx.ext.napoleon'] +
-                                (['sphinx_click'] if poetry.get('scripts') else []))
-exclude_patterns: Final[list[str]] = []
-master_doc: Final[str] = 'index'
-html_static_path: Final[list[str]] = []
-html_theme: Final[str] = 'alabaster'
-templates_path: Final[list[str]] = ['_templates']
+author = f'{authors_list[0]["name"]} <{authors_list[0]["email"]}>'
+copyright = str(datetime.now(timezone.utc).year)  # noqa: A001
+project = name
+release = f'v{version}'
+extensions = [
+    'hoverxref.extension', 'sphinx.ext.autodoc', 'sphinx.ext.intersphinx', 'sphinx.ext.napoleon',
+    'sphinx_datatables', 'sphinx_immaterial', 'sphinxcontrib.autodoc_pydantic',
+    'sphinxcontrib.jquery'
+]
+extensions += ['sphinx_click']
+datatables_class = 'sphinx-datatable'
+datatables_options = {'paging': 0}
+datatables_version = '1.13.4'
+html_theme = 'sphinx_immaterial'
+html_theme_options = {
+    'edit_uri': '/tree/master/docs',
+    'features': [
+        'announce.dismiss', 'content.action.edit', 'content.action.view', 'content.code.copy',
+        'content.tabs.link', 'content.tooltips', 'navigation.expand', 'navigation.footer',
+        'navigation.sections', 'navigation.top', 'search.share', 'search.suggest', 'toc.follow',
+        'toc.sticky'
+    ],
+    'globaltoc_collapse': True,
+    'icon': {
+        'edit': 'material/file-edit-outline',
+        'repo': 'fontawesome/brands/gitlab'
+    },
+    'palette': [{
+        'media': '(prefers-color-scheme)',
+        'toggle': {
+            'icon': 'material/brightness-auto',
+            'name': 'Switch to light mode'
+        }
+    }, {
+        'accent': 'light-blue',
+        'media': '(prefers-color-scheme: light)',
+        'primary': 'light-green',
+        'scheme': 'default',
+        'toggle': {
+            'icon': 'material/lightbulb',
+            'name': 'Switch to dark mode'
+        }
+    }, {
+        'accent': 'lime',
+        'media': '(prefers-color-scheme: dark)',
+        'primary': 'deep-orange',
+        'scheme': 'slate',
+        'toggle': {
+            'icon': 'material/lightbulb-outline',
+            'name': 'Switch to system preference'
+        }
+    }],
+    'repo_name': 'usps-track',
+    'repo_url': 'https://github.com/Tatsh/usps-track',
+    'site_url': 'https://usps-track.readthedocs.org',
+    'toc_title_is_page_title': True
+}
+intersphinx_mapping = {'python': ('https://docs.python.org/3', None)}
