@@ -1,12 +1,17 @@
-from typing import Any, Iterable, TypedDict, cast
+from collections.abc import Iterable
+from typing import TypedDict, cast
 import logging
 
 from aiohttp import TraceConfig
 from aiohttp.client_exceptions import ContentTypeError
 import aiohttp
 
-from .constants import (DEFAULT_STR_VALUE, HEADERS, URL_TRACK_CONFIRM_ACTION,
-                        URL_TRACK_CONFIRM_UPDATE)
+from .constants import (
+    DEFAULT_STR_VALUE,
+    HEADERS,
+    URL_TRACK_CONFIRM_ACTION,
+    URL_TRACK_CONFIRM_UPDATE,
+)
 from .utils import on_off
 
 __all__ = ('TextServiceError', 'usps_track')
@@ -25,12 +30,23 @@ class TextServiceError(Exception):
 async def usps_track(phone_number: str,
                      tracking_numbers: Iterable[str],
                      trace_configs: list[TraceConfig] | None = None,
-                     raise_for_status: bool = False,
-                     **kwargs: Any) -> None:
+                     *,
+                     confirm_sms: bool = True,
+                     email: str = DEFAULT_STR_VALUE,
+                     name: str = DEFAULT_STR_VALUE,
+                     text_alert: bool = True,
+                     text_all: bool = True,
+                     text_dnd: bool = True,
+                     text_future: bool = True,
+                     text_oa: bool = True,
+                     text_pickup: bool = True,
+                     text_today: bool = True,
+                     raise_for_status: bool = False) -> None:
     """
+    Track a package using the USPS tracking system via SMS.
+
     Parameters
     ----------
-
     phone_number : str
         Must be a US phone number including area code without a leading '1-', and with each group
         separated using '-' as the delimiter.
@@ -68,35 +84,35 @@ async def usps_track(phone_number: str,
     async with aiohttp.ClientSession(headers=HEADERS, trace_configs=trace_configs) as session:
         for number in tracking_numbers:
             async with session.get(URL_TRACK_CONFIRM_ACTION,
-                                   params=dict(qtc_tLabels1=number),
+                                   params={'qtc_tLabels1': number},
                                    raise_for_status=raise_for_status) as r:
                 if 'could not locate the tracking information' in await r.text():
                     continue
                 async with session.post(URL_TRACK_CONFIRM_UPDATE,
-                                        data=dict(
-                                            confirmSms=on_off(kwargs.get('confirm_sms', True)),
-                                            email1=kwargs.get('email', DEFAULT_STR_VALUE),
-                                            label=number,
-                                            name1=kwargs.get('name', DEFAULT_STR_VALUE),
-                                            smsNumber=phone_number,
-                                            textAlert=on_off(kwargs.get('text_alert', True)),
-                                            textAll=on_off(kwargs.get('text_all', True)),
-                                            textDnd=on_off(kwargs.get('text_dnd', True)),
-                                            textFuture=on_off(kwargs.get('text_future', True)),
-                                            textOA=on_off(kwargs.get('text_oa', True)),
-                                            textPickup=on_off(kwargs.get('text_pickup', True)),
-                                            textToday=on_off(kwargs.get('text_today', True)),
-                                        ),
+                                        data={
+                                            'confirmSms': on_off(confirm_sms),
+                                            'email1': email,
+                                            'label': number,
+                                            'name1': name,
+                                            'smsNumber': phone_number,
+                                            'textAlert': on_off(text_alert),
+                                            'textAll': on_off(text_all),
+                                            'textDnd': on_off(text_dnd),
+                                            'textFuture': on_off(text_future),
+                                            'textOA': on_off(text_oa),
+                                            'textPickup': on_off(text_pickup),
+                                            'textToday': on_off(text_today)
+                                        },
                                         headers={
                                             'Referer':
                                                 f'{URL_TRACK_CONFIRM_ACTION}?qtc_tLabels1={number}',
                                             'X-Requested-With':
-                                                'XMLHttpRequest',
+                                                'XMLHttpRequest'
                                         },
                                         raise_for_status=raise_for_status) as r2:
                     try:
-                        if cast(ResponseDict, await r2.json()).get('textServiceError') != 'false':
-                            raise TextServiceError()
+                        if cast('ResponseDict', await r2.json()).get('textServiceError') != 'false':
+                            raise TextServiceError
                     except ContentTypeError:
                         logger.debug('Error updating %s. Tracking number probably does not exist.',
                                      number)
